pyproject.toml
[project]
name = "dcloak"
version = "0.1.0"
description = ""
authors = [
    {name = "Jelmer de Reus",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "django-dotenv (==1.4.2)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-allauth[idp-oidc] (>=65.11.2,<66.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "cryptography (>=45.0.7,<46.0.0)"
]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
